refactor: remove date cell and empty placeholder
